# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Messenger backend api",
        "description": "Messenger backend",
        "contact": {
            "email": "tcom.developer@gmail.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://crypt-talk.test/api",
            "description": "Dev server"
        },
        {
            "url": "http://95.217.156.195/api",
            "description": "Dev server"
        }
    ],
    "paths": {
...
        "/auth/register": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Register new user",
                "operationId": "ed42f392010c301892e640067141676e",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "name": {
                                        "description": "user name",
                                        "type": "string",
                                        "default": "newuser"
                                    },
                                    "email": {
                                        "description": "email",
                                        "type": "string",
                                        "default": "newuser@mail.com"
                                    },
                                    "password": {
                                        "description": "password",
                                        "type": "string",
                                        "default": "Secret123#"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "User registered"
                    },
                    "500": {
                        "description": "Server Error"
                    }
                }
            }
        },
        "/auth/resend-verify-email": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "resend verify email",
                "operationId": "39227350a2b92d59775b309975bce2d5",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email"
                                ],
                                "properties": {
                                    "email": {
                                        "description": "email",
                                        "type": "string",
                                        "default": "resend@mail.com"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Verify email sent"
                    },
                    "500": {
                        "description": "Server Error"
                    }
                }
            }
        },
        "/auth/verify-email": {
            "get": {
                "tags": [
                    "Auth"
                ],
                "summary": "Verify email address",
                "operationId": "6902a4b313dbb202a9d4592dc5acd616",
                "parameters": [
                    {
                        "name": "verify_token",
                        "in": "query",
                        "description": "Email verification token from email link",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9..."
                    }
                ],
                "responses": {
                    "302": {
                        "description": "Redirects to frontend login page"
                    },
                    "400": {
                        "description": "Invalid or expired token"
                    },
                    "404": {
                        "description": "User not found"
                    }
                }
            }
        },
        "/auth/login": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Login user",
                "operationId": "8cb7742ae98d0990acb2907c4e00bf24",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "description": "email",
                                        "type": "string",
                                        "default": "milos@mail.com"
                                    },
                                    "password": {
                                        "description": "password",
                                        "type": "string",
                                        "default": "milos"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Korisnik prijavljen"
                    },
                    "500": {
                        "description": "Server Error"
                    }
                }
            }
        },
        "/auth/refresh": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Generate new token pair",
                "operationId": "f6c9683ed975ace7d035a09e6e2a9332",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "refresh_token"
                                ],
                                "properties": {
                                    "refresh_token": {
                                        "description": "refresh_token",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Token pair created"
                    },
                    "500": {
                        "description": "Server Error"
                    }
                }
            }
        },
        "/auth/whoami": {
            "get": {
                "tags": [
                    "Auth"
                ],
                "summary": "Podaci o korisniku na osnovu tokena",
                "operationId": "aa7bc66822f59c40cade93c5e4c3b41e",
                "responses": {
                    "200": {
                        "description": "Podaci o korisniku vraćeni na front"
                    },
                    "500": {
                        "description": "Server Error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/auth/logout": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Odjava sa sistema",
                "operationId": "0d29c740dc06d77602af8095bf4654cc",
                "responses": {
                    "200": {
                        "description": "Korisnik odjavljen iz sistema"
                    },
                    "500": {
                        "description": "Server Error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/auth/forgot-password": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Send reset password to email",
                "operationId": "b13d5efc90e1de6d4e9900240fce9b4d",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email"
                                ],
                                "properties": {
                                    "email": {
                                        "description": "User email",
                                        "type": "string",
                                        "example": "milos@mail.com"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Reset password link sent"
                    },
                    "500": {
                        "description": "Server Error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/auth/profile-update": {
            "post": {
                "tags": [
                    "User"
                ],
                "summary": "User profil update",
                "operationId": "4a6525e8e9aadee174355e6371e78f7b",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "name",
                                    "about",
                                    "avatar"
                                ],
                                "properties": {
                                    "name": {
                                        "description": "User name",
                                        "type": "string",
                                        "example": "John Doe"
                                    },
                                    "about": {
                                        "description": "About user",
                                        "type": "string",
                                        "example": "Software developer from Belgrade"
                                    },
                                    "avatar": {
                                        "description": "User avatar file",
                                        "type": "string",
                                        "format": "binary"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "User updated"
                    },
                    "500": {
                        "description": "Server Error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/user/password-update": {
            "post": {
                "tags": [
                    "User"
                ],
                "summary": "User password update",
                "operationId": "6e8605d5e28946b6e118a8db9a85469d",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "name",
                                    "about",
                                    "avatar"
                                ],
                                "properties": {
                                    "name": {
                                        "description": "User name",
                                        "type": "string",
                                        "example": "John Doe"
                                    },
                                    "about": {
                                        "description": "About user",
                                        "type": "string",
                                        "example": "Software developer from Belgrade"
                                    },
                                    "avatar": {
                                        "description": "User avatar file",
                                        "type": "string",
                                        "format": "binary"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Password updated"
                    },
                    "500": {
                        "description": "Server Error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Authorize with bearer token",
                "name": "Authorization",
                "in": "header",
                "scheme": "Bearer"
            }
        }
    },
...
}